manager

+++ src/modules/validator/validatorService.ts
@@ -25,6 +25,25 @@ export class ValidatorService {
     data: { userId: number; rank: number }
   ) => {
     const count = (await this.getValidatorByCategory(id)).length;
+    // find manager Id
+    const manager = await prisma.role.findFirst({
+      where: {
+        label: "MANAGER",
+      },
+    });
+
+    if (!manager)
+      throw Error("MANAGER Role not created, please create the role first");
+
+    await prisma.user.update({
+      where: { id: data.userId },
+      data: {
+        role: {
+          connect: { id: manager.id },
+        },
+      },
+    });
+
     if (count <= 3) {
       return prisma.validator.create({
         data: {
@@ -39,11 +58,31 @@ export class ValidatorService {
   removeValidatorFromCategory = async (id: number) => {
     const count = (await this.getValidatorByCategory(id)).length;
     if (count > 0) {
-      return prisma.validator.delete({
+      // find manager Id
+      const manager = await prisma.role.findFirst({
+        where: {
+          label: "MANAGER",
+        },
+      });
+
+      if (!manager) throw Error("MANAGER Role not created");
+
+      const validator = await prisma.validator.delete({
         where: {
           id,
         },
       });
+
+      await prisma.user.update({
+        where: { id: validator.userId },
+        data: {
+          role: {
+            disconnect: { id: manager.id },
+          },
+        },
+      });
+
+      return validator;
     }
     throw Error("No validator in this category");
   };
